Move offline indicator to topbar
